be/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langchain-community==0.3.26
langchain-qdrant==0.2.0
langsmith==0.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.97.1
orjson==3.11.0
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
sentry-sdk==2.33.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
zstandard==0.23.0
